x64 warning fixes.

+++ objmdl/ObjectCollection.cpp
@@ -33,7 +33,7 @@ STDMETHODIMP CObjectCollection::RemoveObjectAt(UINT uiIndex)
 STDMETHODIMP CObjectCollection::GetCount(UINT *pcObjects)
 {
 	CHECK_E_POINTER(pcObjects);
-	*pcObjects = m_objects.size();
+	*pcObjects = static_cast<UINT>(m_objects.size());
 	return S_OK;
 }
 
@@ -60,6 +60,6 @@ STDMETHODIMP CObjectCollection::IndexOf(void* pv, UINT* puiIndex)
 	if (it == m_objects.end())
 		return E_NOT_SET;
 
-	*puiIndex = std::distance(m_objects.begin(), it);
+	*puiIndex = static_cast<UINT>(std::distance(m_objects.begin(), it));
 	return S_OK;
 }
